Revert "HBASE-19835 Use explicit casting to avoid side effects"
This reverts commit bc080e75000aa16ee05d42033ec5e4d984985205.

 Conflicts:
	hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/HStore.java
	hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/Segment.java

patch reverted changes that happened in parallel without explanation. see jira.

+++ hbase-client/src/main/java/org/apache/hadoop/hbase/client/Mutation.java
@@ -66,7 +66,7 @@ public abstract class Mutation extends OperationWithAttributes implements Row, C
     HeapSize {
   public static final long MUTATION_OVERHEAD = ClassSize.align(
       // This
-      (long)ClassSize.OBJECT +
+      ClassSize.OBJECT +
       // row + OperationWithAttributes.attributes
       2 * ClassSize.REFERENCE +
       // Timestamp
@@ -791,10 +791,10 @@ public abstract class Mutation extends OperationWithAttributes implements Row, C
 
   private static final class CellWrapper implements ExtendedCell {
     private static final long FIXED_OVERHEAD = ClassSize.align(
-          (long)ClassSize.OBJECT              // object header
+      ClassSize.OBJECT              // object header
         + KeyValue.TIMESTAMP_SIZE       // timestamp
         + Bytes.SIZEOF_LONG             // sequence id
-        + 1L * ClassSize.REFERENCE);     // references to cell
+        + 1 * ClassSize.REFERENCE);     // references to cell
     private final Cell cell;
     private long sequenceId;
     private long timestamp;
